Move config file loading to more appropriate packages.
